页面组件整合和封装 蹇坤霖 #140
