+++ src/config/router.config.js
@@ -113,14 +113,12 @@ export const asyncRouterMap = [
           component: () => import('@/views/stat/StatOverview'),
           meta: { title: '统计概览', permission: [ 'dashboard' ] }
         },
-        // chenbin_modify
         {
           path: '/stat/courseanalysis', // todo
           name: '课程分析',
           component: () => import('@/views/stat/StatCourseAnalysis'),
           meta: { title: '课程分析', permission: [ 'dashboard' ] }
         },
-        // chenbin_modify
         {
           path: '/stat/statstudentdata',
           name: '学生统计',
